refactor(model-selector): drop llmProviders prop; fetch via useCurrentAgent internally
ModelSelector, ModelSelectorContent, and MultiModelSelector now resolve
LLM providers internally by calling useLLMProviders(useCurrentAgent()?.id).
In chat contexts this gives persona-scoped providers; on admin/editor pages
where no agent is active it falls back to user-scoped providers.

Also moves useLanguageModels.ts → lib/languageModels/hooks.ts to co-locate
it with the rest of the languageModels library, and updates all references.
