+++ web/src/app/craft/onboarding/hooks/useOnboardingModal.ts
@@ -2,7 +2,7 @@
 
 import { useCallback, useState, useMemo, useEffect } from "react";
 import { useUser } from "@/providers/UserProvider";
-import { useLLMProviders } from "@/hooks/useLanguageModels";
+import { useLLMProviders } from "@/lib/languageModels/hooks";
 import {
   OnboardingModalMode,
   OnboardingModalController,
